Change JPA repository ID type from Integer to Long
Updated all repository interfaces to use Long as the ID type instead of Integer for consistency and scalability. Also removed misplaced @Id annotation from MaterialUrl entity.

+++ src/main/java/com/solarize/solarizeWebBackend/modules/budget/BudgetRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BudgetRepository extends JpaRepository<Budget, Integer> {
+public interface BudgetRepository extends JpaRepository<Budget, Long> {
 }
